libs: Bump gplyapi to 3.5.0
* Map versionCode as Long
* Handle all cases of immutable object modification

+++ app/src/main/java/com/aurora/store/data/helper/DownloadHelper.kt
@@ -107,7 +107,7 @@ class DownloadHelper @Inject constructor(
      * @param packageName Name of the package of the app
      * @param versionCode Version of the package
      */
-    suspend fun clearDownload(packageName: String, versionCode: Int) {
+    suspend fun clearDownload(packageName: String, versionCode: Long) {
         Log.i(TAG, "Clearing downloads for $packageName ($versionCode)")
         downloadDao.delete(packageName)
         PathUtil.getAppDownloadDir(context, packageName, versionCode)
