+++ app/src/main/java/com/aurora/store/data/installer/ShizukuInstaller.kt
@@ -129,7 +129,7 @@ class ShizukuInstaller @Inject constructor(
 
     private fun install(
         packageName: String,
-        versionCode: Int,
+        versionCode: Long,
         sharedLibPkgName: String = "",
         displayName: String = ""
     ) {
